project1/backend/requirements.txt
bcrypt==3.2.0
#Flask==1.1.2
Flask==2.2.2
Flask-API==3.0.post1
Flask-Cors==3.0.9
jsons==1.3.0
pycryptodome==3.9.9
pytest==6.2.1